requirements.txt
django
django-tastypie
django-crontab==0.7.1
Django==3.2.3
djangorestframework==3.12.4
python-decouple==3.4
psycopg2==2.8.6
dj-static==0.0.6
certifi
requests==2.25.1
pyyaml
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
uritemplate
PyYAML
django-cors-headers
pandas
numpy
wordcloud
matplotlib
scikit-learn
plotly
